requirements.txt
fastapi==0.115.12
passlib==1.7.4
pytest==8.3.5
python-dotenv==1.1.0
shortuuid==1.0.13
SQLAlchemy==2.0.40
traitlets==5.14.3
